Undoing the change where I added DatabaseUtil as I ended up not needing to add another database upgrade task.
